# api/openapi.yaml
openapi: 3.0.3
info:
  title: OpenDash API
  version: 0.1.0
  description: >-
    Persistent home-server dashboard API with single-administrator authentication.
    When authentication is disabled, dashboard routes are public, bootstrap reports
    authEnabled=false/bootstrapRequired=false, and all other auth endpoints return 404.
security:
  - sessionCookie: []
servers:
  - url: http://127.0.0.1:8080/api/v1
    description: Local development server
paths:
...
  /catalog/{id}/install:
    post:
      summary: Install a catalog app
      operationId: installApp
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                config:
                  type: object
                  additionalProperties:
                    type: string
      responses:
        "202":
          description: Install operation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "409":
          description: App already installed, operation in progress, or conflicts
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: opendash_session
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
  schemas:
...
    Operation:
      type: object
      required:
        - id
        - appId
        - kind
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        appId:
          type: string
        kind:
          type: string
        status:
          type: string
        error:
          type: string
        output:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...